No printing of repeated values, see below for interesting observations
So the first thing I did was to stop it from serial printing a value if
it was the exact same thing as before. This worked fine for digital. For
analog though, there was a problem. It just started printing out values
that were 1 or 2 milliamps or whatever the unit is above the last value
with no change in the sensor. No matter, I thought. I'll just increase
the tollerance to 5. Mysteriously, it now printed out values that were 5
or more above or below the last value. Now came my observation. You see
young grasshopper, the arduino was making 1000s of measurements a
second. If it did not detect a change, it just made another measurement.
The measurements probably followed a bell shaped curve. So by the law of
large numbers numbers any distance away from the mean would eventually
get picked given enough measurements. By increasing the tolerance, I got
the arduino to just print the outliers, but there was still no change in
the sensor. I had to make the tolerance the high value of 50 for most of
the wrong values to appear. This has the side effect of making any
actual sensor value less than 50 away from the current one not getting
picked up. This deservers further thought on how to best decide which
measurments to send to the ESP to send to the computer.